Moved getToolByName early patch to the later patches.
This fixes a circular import.
See https://github.com/plone/Products.CMFPlone/issues/1950
